Delete related interfaces when stopping a role

+++ ap/ap_stop.sh
@@ -3,9 +3,19 @@ ap_pid=$(<"/var/run/hostapd.pid")
 
 echo "Terminating 1st AP - pid " $ap_pid
 
+if [ -n "$ap_pid" ]; then
 kill $ap_pid
+fi
 
 #ps | grep -v udhcpd2 | grep u[d]hcpd
 output=`ps | grep -v udhcpd2 | grep u[d]hcpd`
 set -- $output
-kill $1
+
+if [ -n "$output" ]; then
+ kill $1
+fi
+
+if [ -d /sys/class/net/wlan1 ]
+then
+ iw wlan1 del
+fi
